Update pre-commit configuration and CI workflow
This updates `black` to the last version that supported Python 2
formatting (including a pin to its `click` dependency) and updates the
URL of the flake8 repository which has been moved to GitHub.

It also uses `ubuntu-20.04` as the job runner in the GitHub test
workflow because Python 3.6 is not supported in `ubuntu-22.04`
anymore.